Eliminate CommandBuilder class
