+++ src/NUnitFramework/framework/Internal/Execution/CompositeWorkItem.cs
@@ -83,11 +83,6 @@ namespace NUnit.Framework.Internal.Execution
         /// </summary>
         protected override void PerformWork()
         {
-            // Initialize actions, setup and teardown
-            // We can't do this in the constructor because
-            // the context is not available at that point.
-            InitializeSetUpAndTearDownCommands();
-
             if (!CheckForCancellation())
                 if (Test.RunState == RunState.Explicit && !Filter.IsExplicitMatch(Test))
                     SkipFixture(ResultState.Explicit, GetSkipReason(), null);
@@ -105,6 +100,8 @@ namespace NUnit.Framework.Internal.Execution
 
                             if (_children.Count > 0)
                             {
+                                InitializeSetUpAndTearDownCommands();
+
                                 PerformOneTimeSetUp();
 
                                 if (!CheckForCancellation())
@@ -169,9 +166,8 @@ namespace NUnit.Framework.Internal.Execution
 
         private void InitializeSetUpAndTearDownCommands()
         {
-            List<SetUpTearDownItem> setUpTearDownItems = _suite.TypeInfo != null
-                ? CommandBuilder.BuildSetUpTearDownList(_suite.TypeInfo.Type, typeof(OneTimeSetUpAttribute), typeof(OneTimeTearDownAttribute))
-                : new List<SetUpTearDownItem>();
+            List<SetUpTearDownItem> setUpTearDownItems = 
+                BuildSetUpTearDownList(typeof(OneTimeSetUpAttribute), typeof(OneTimeTearDownAttribute));
 
             var actionItems = new List<TestActionItem>();
             foreach (ITestAction action in Test.Actions)
@@ -188,10 +184,10 @@ namespace NUnit.Framework.Internal.Execution
                 // ParameterizedMethodSuites and individual test cases both use the same
                 // MethodInfo as a source of attributes. We handle the Test and Default targets
                 // in the test case, so we don't want to doubly handle it here.
-                bool applyToSuite = (action.Targets & ActionTargets.Suite) == ActionTargets.Suite
+                bool applyToSuite =  action.Targets.HasFlag(ActionTargets.Suite)
                     || action.Targets == ActionTargets.Default && !(Test is ParameterizedMethodSuite);
 
-                bool applyToTest = (action.Targets & ActionTargets.Test) == ActionTargets.Test
+                bool applyToTest = action.Targets.HasFlag(ActionTargets.Test)
                     && !(Test is ParameterizedMethodSuite);
 
                 if (applyToSuite)
@@ -201,8 +197,60 @@ namespace NUnit.Framework.Internal.Execution
                     Context.UpstreamActions.Add(action);
             }
 
-            _setupCommand = CommandBuilder.MakeOneTimeSetUpCommand(_suite, setUpTearDownItems, actionItems);
-            _teardownCommand = CommandBuilder.MakeOneTimeTearDownCommand(_suite, setUpTearDownItems, actionItems);
+            _setupCommand = MakeOneTimeSetUpCommand(setUpTearDownItems, actionItems);
+
+            _teardownCommand = MakeOneTimeTearDownCommand(setUpTearDownItems, actionItems);
+        }
+
+        private TestCommand MakeOneTimeSetUpCommand(List<SetUpTearDownItem> setUpTearDown, List<TestActionItem> actions)
+        {
+            TestCommand command = new EmptyTestCommand(Test);
+
+            // Add Action Commands 
+            int index = actions.Count;
+            while (--index >= 0)
+                command = new BeforeTestActionCommand(command, actions[index]);
+
+            if (Test.TypeInfo != null)
+            {
+                // Build the OneTimeSetUpCommands
+                foreach (SetUpTearDownItem item in setUpTearDown)
+                    command = new OneTimeSetUpCommand(command, item);
+
+                // Construct the fixture if necessary
+                if (!Test.TypeInfo.IsStaticClass)
+                    command = new ConstructFixtureCommand(command);
+            }
+
+            // Prefix with any IApplyToContext items from attributes
+            foreach (var attr in Test.GetCustomAttributes<IApplyToContext>(true))
+                command = new ApplyChangesToContextCommand(command, attr);
+
+            return command;
+        }
+
+        private TestCommand MakeOneTimeTearDownCommand(List<SetUpTearDownItem> setUpTearDownItems, List<TestActionItem> actions)
+        {
+            TestCommand command = new EmptyTestCommand(Test);
+
+            // Dispose of fixture if necessary
+            if (Test is IDisposableFixture && typeof(IDisposable).IsAssignableFrom(Test.TypeInfo.Type))
+                command = new DisposeFixtureCommand(command);
+
+            // Create the OneTimeTearDown commands
+            foreach (SetUpTearDownItem item in setUpTearDownItems)
+                command = new OneTimeTearDownCommand(command, item);
+
+            // Create the AfterTestAction commands
+            int index = actions.Count;
+            while (--index >= 0)
+                command = new AfterTestActionCommand(command, actions[index]);
+
+            // For Theories, follow with TheoryResultCommand to adjust result as needed
+            if (Test.TestType == "Theory")
+                command = new TheoryResultCommand(command);
+
+            return command;
         }
 
         private void PerformOneTimeSetUp()
