+++ src/NUnitFramework/framework/nunit.framework-4.5.csproj
@@ -342,7 +342,6 @@
     <Compile Include="Internal\InvalidDataSourceException.cs" />
     <Compile Include="Internal\CultureDetector.cs" />
     <Compile Include="Internal\ExceptionHelper.cs" />
-    <Compile Include="Internal\Execution\CommandBuilder.cs" />
     <Compile Include="Internal\Execution\CompositeWorkItem.cs" />
     <Compile Include="Internal\Execution\CountdownEvent.cs" />
     <Compile Include="Internal\Execution\EventPump.cs" />
